Revert helper change to fix broken acceptance test
https://eaflood.atlassian.net/browse/WATER-4992

In [Add fetch notice service and unit tests](https://github.com/DEFRA/water-abstraction-system/pull/1967) we updated the `scheduled-notification.helper.js` to include a bunch more default values.

The problem is one of them was to set an `eventId` but the table has a foreign key constraint that the value must match an existing `water.events`.

The unit test we think the changes were made for, actually only need a few of the fields to be set. We didn't need to default the majority of them for all the tests to still pass.

Plus, we should have caught it at review, but this breaks the idea of the 'defaults'. They should only be values that are needed for the DB to allow us to create a record without error. All other values should come from the unit tests themselves. Hence, the helper prior to the change was only setting the `createdAt` value.

Anyway, the change meant when the acceptance tests data loader was trying to add a scheduled notification using the helper, it was failing because of an unknown `eventId`.

This reverts the change to the helper, which brings it back inline with our conventions and gets the acceptance test passing again.