Revert helper change to fix broken acceptance test (#1976)
https://eaflood.atlassian.net/browse/WATER-4992

In [Add fetch notice service and unit tests](https://github.com/DEFRA/water-abstraction-system/pull/1967) we updated the `scheduled-notification.helper.js` to include a bunch more default values.

The problem is that one was to set an `eventId`, but the table has a foreign key constraint: the value must match an existing `water.events`.

The unit test we think the changes were made for only needs a few fields to be set. We didn't need to default most of them for all the tests to pass.

We should have caught it at review, but this breaks the idea of the 'defaults'. They should only be values needed by the DB to allow us to create a record without error. All other values should come from the unit tests themselves. Hence, before the change, the helper only set the `createdAt` value.

Anyway, the change meant that when the acceptance tests data loader tried to add a scheduled notification using the helper, it failed because of an unknown `eventId`.

This reverts the change to the helper, bringing it back in line with our conventions and gets the acceptance tests passing again.