Create `GenerateReturnSubmissionMetadataService` (#1960)
https://eaflood.atlassian.net/browse/WATER-4847

Broken out of PR https://github.com/DEFRA/water-abstraction-system/pull/1815

This PR implements `GenerateReturnSubmissionMetadataService`. The service generates the object that will be persisted in the `metadata` field of the return submission, using the data passed in the session.

If this is a nil return (which we determine based on the user's current journey as stored in `session.journey`) then no metadata is needed so the service returns an empty object.

Otherwise, we format the metadata object in the expected format and return it.