Feature notificaitons setup abstraction alert check licence matches view table (#1974)
https://eaflood.atlassian.net/browse/WATER-5027

The legacy 'abstraction-alerts/check-licence-matches' page has a table displaying the selected thresholds.

We already have a table for the monitoring stations page, we can use this table in multiple places.

This change adds / updates the logic to display the monitoring restrictions for the check licence matches page.

The table will only show the previously selected

This change does not implement the remove link logic, but the link and text have been implemented to simplify development.