Migration - Add `created_by` to `licence_monitoring_stations` (#1977)
https://eaflood.atlassian.net/browse/WATER-4999

As part of the work migrating the "Monitoring stations - Tag details page" we are now required to capture the user that created the Tag.

This PR will create a `created_by` column in the `public.licence_monitoring_stations` view and legacy table to enable us to capture the user ID.

The models will also be updated to add the relationship to the `users`.