Update notifications setup abstraction alert types error message
https://eaflood.atlassian.net/browse/WATER-5033

When a user selects an alert type that is not warning or resume. We need to check that the selected alert type is available as a restriction type to send the alert notification.

This change updates the alert type submit service and validator to check if the selected alert type is available.