Update the FetchNoticesService
This involved 3 strands of work

- fix errors in the query
- remove mapping logic
- refactor to make steps distinct

On the error front, the query being generated was incorrect in two ways.

1 - It couldn't handle both an alert notice and a standard notice type being selected. Adding the alert type adds the `whereRaw()` that looks in the metadata field for an alert type. But standard notices don't have this property, which means those records would never match. In effect, selecting any alert notice type _automatically_ excludes all standard notices.

This explains why the `_applyNoticeTypeFilters()` needs a special clause to handle when both are selected.

2 - The other issue is that it was looking at the wrong alert type flag. Checked with the data and as a side-by-side comparison. The `'options'->'linkages'->0->0->>'alertType` it was checking comes from the 'restriction' record the alert notice is based on. They are 'stop', 'stop or reduce', or 'reduce'. For example, we might send a 'warning' or a 'resume' notice for a licence with a 'stop' restriction.

So, it is `'options'->>'sendingAlertType'` we need to be retrieving and filtering against.

On the refactor front, there is something that will pertain to all the modules and then just general stuff.

The 'global' refactor is I want to get away from each module having to map values from one thing to another. Ideally, the values selected in the view are the values we use in the queries and thus no mapping is needed. So, all that has been stripped out.

The rest was just breaking down the methods as IMHO they were too big and doing to much. Hopefully, they now have single jobs and things are a little clearer to follow.