Missing validation for quarterly return & summer combination (#1983)
* Missing validation for quarterly return & summer combination

https://eaflood.atlassian.net/browse/WATER-5052

During [Amend validation for returns marked quarterly submissions](https://github.com/DEFRA/water-abstraction-system/pull/1903), a fix was added to prevent requirements for returns with a
combination of summer and marked quarterly return submission. While being tested by the QA team, it was noticed that the current implementation does not take into account licences marked as water
undertakers.

When a licence is an undertaker, the setup automatically sets quarterly return submissions to true, which allows invalid data to be submitted. In order to fix this, when a licence is a water undertaker
the user is prevented from selecting `summer` on the returns cycle page.

This PR adds additional validation to the return cycle page to prevent licences marked as water undertakers from submitting a combination of summer return cycle and quarterly return submission.