Add `console.dir()` check to github CI
https://github.com/DEFRA/water-abstraction-team/issues/136

Previously, we had an issue with `console.log()` being found in the system code base - [Check for console.log() in CI](https://github.com/DEFRA/water-abstraction-team/issues/94).
This was resolved by adding a check to the Github CI prevent successful builds when `console.log()` statements are present in the codebase.

Now with the use of `console.dir()` to print full Javascript objects, thanks to [Logify 🚀 ](https://github.com/rvsiyad/logify), there has been similar instances of forgetting to
remove the print statements, which the CI will not pick up due to the different type of logging. This has led to changes being requested to PRs which has become repetitive. To avoid this,
it would be great if our CI could catch these in the future to prevent us from making this mistake.

This PR adds a check for `console.dir()` to the CI.