Add notification setup abstraction alert email scaffold
Users should be able to select an email address which will receive the abstraction alerts. This can either be the
email address linked to the licence or another email selected by the user. The abstraction alert email page will
give users the choice between using the licence linked email address or another email address that must be inputted.

This PR adds the alert email address view for the notification setup journey.