Update notifications setup abstraction alert cancel
https://eaflood.atlassian.net/browse/WATER-5056

We have a pattern on check pages to allow the user to cancel the journey.

This change updates the 'cancel-alert' services and presenter to show the cancel page and allow the user to continue cancelling.

This will delete the session and redirect the user to the page prior to the start of the journey.