Update notices setup abstraction alerts - alert type (#1993)
* Update notices setup abstraction alerts - alert type

https://eaflood.atlassian.net/browse/WATER-5014

We have introduced an additional validation step which is triggered when there are no licence monitoring stations with the alert type the user has selected.

When we added this change we missed setting the selected box with the users provided alert type.

This change updates the alert-type submit service to pass the user selected alert type into the alert type submit presenter.