Update logic for checking which return logs were created (#1994)
https://eaflood.atlassian.net/browse/WATER-4547

As part of our prep to go live with quarterly returns the annual return log job was ran on pre-prod with the latest data. As part of the testing our testers saw an issue with return logs not being voided for partial replacements of a return version. It works as expected if you replace the entire return version but not when you create one that starts part way through the previous one.

The logic of the creation of return logs keeps track of all the return logs that is has created and then voids any that were not created during this process. This works fine if you're the latest return version but if you're an interstitial one it causes all of the future ones to be voided.

This PR does two things:

It uses Knex to return the actually inserted new records which are then used to filter out what needs voiding. We can't use objectionjs due to an open bug around reporting onConflict: Vincit/objection.js#2320

It also checks for the end date of the return version you're creating so that it doesn't void beyond what is necessary