Add notifications setup abstraction check licence submit service
https://eaflood.atlassian.net/browse/WATER-5061

We have existing check recipient functionality for sending notification.

To use this functionality we need to create a list of licence refs for the recipient fetch and download service to use.

Ths changes adds the check licence submit service, which will redirect to the email address page. It will also create an array of licence refs for the recipient page to use in its query.

Doing this here save putting any logic in the already complicated fetch-recipients.service.