Add notifications setup abstraction alert to recipients
https://eaflood.atlassian.net/browse/WATER-5061

We have existing check recipient functionality for sending notification.

To use this functionality for abstraction alerts we previously created a list of licence refs for the recipient fetch and download service to use.

Ths changes adds the abstraction alerts recipients logic to the recipients service.

Changes have been made around