Update Two-part tariff annual years for 2024-25
https://eaflood.atlassian.net/browse/WATER-5064

Currently, when you create a two-part tariff annual bill run, the journey presents you with a screen where you pick the financial year for which you are creating it.

This is because, for 'reasons', a backlog of two-part tariff annual bill runs has built up. Ordinarily, they would be processed yearly, so they it behaves like standard annual. Users don't select a year because the 'current' year (and for 2PT, 'current' means the previous financial year) is assumed.

Knowing this page in the annual year is only temporary, we decided to hard-code the years presented rather than waste resources trying to be 'clever'.

This ticket would have been to the 'drop the page'; however, some of the older two-part tariff bill runs are still on the backlog to be processed. Only 2022-23 and 2023-24 have been.

This change updates the page to display the following options

- 2024 to 2025
- 2021 to 2022
- 2020 to 2021