Update the IndexNoticesPresenter
This covers 3 strands of work

- fix the errors in the output
- refactor to remove presentation code from the presenter
- properly support pagination details

On the error front, the presenter was only prefixing the alert type to warnings. If it was a reduce, resume or stop abstraction alert it was only outputting `Water abstraction alert` on the page.

Regards the refactoring, we're trying to keep a distinction between logic that prepares the data ready for display, and how it is displayed.

Presenters are responsible for taking data fetched from the db and either formatting or transforming it into what is needed for display. But they are not responsible for _how_ the data is displayed, for example, is it bold, or a status tag, or to be used in a link.

That is the responsibility of the view. On a similar note, there is no formatting needed for the table headers, nor logic to fetch or determine them. Therefore it is content which is perfectly fine to live in the view.

The final changes were around pagination support. If we have more than one page of results we include this detail in the page title (take a look at bill runs). Therefore you need the number of results plus how many were displaying. Because these details come from the PaginationPresenter, we work out the page title in the service rather than the presenter.

What we _do_ need to do though is properly set the table caption (check out the table in review two-part tariff bill run). So, we need the number we're displaying and the total number. Add to that the number of notices we'll actually be dealing with in `prod` means we have to format the number to be human readable.