Fix meter readings not correctly being persisted on submission
https://eaflood.atlassian.net/browse/WATER-4847

During QA it was found that volumes were not being persisted for meter readings. On investigation, it was found that although we calculate volumes from meter readings for display, these calculated volumes aren't put in the session. This was overlooked as when the session is initialised from a previous return, we do populate it with volumes.

We fix this by ensuring that when the user submits a return, we re-calculate volumes based on meter readings prior to persisting the lines.