Update billing account view to use new change address endpoint (#2005)
https://eaflood.atlassian.net/browse/WATER-4988

During testing, it was noticed that the change address button for the new billing account page did not work. This turned out to be because the endpoint used by the legacy page would redirect to another endpoint, which also required session data to work. To get around this issue, we [incorporated a new endpoint](https://github.com/DEFRA/water-abstraction-ui/pull/2693) that can be hit from the new system to reach the legacy change address page. We now need to update the view billing account change address button to use the new endpoint.

This PR updates the view billing account change address link to the new endpoint.