Fix billing account page to only show sent bills
https://eaflood.atlassian.net/browse/WATER-4988

While testing the billing account page, it was noticed that some of the bills being displayed were not sent.
This means bills that have a status of `error`, `ready` and `review` are also being shown. To amend this, we must only
show bills that have a status of `sent`.

This PR will fix the billing account page to only show `sent` bills.