Update the IndexService
Now we're getting into fundamental changes. Prior to this the validation of what had been submitted was done here. The results of the validation would then determine how the page was displayed.

This breaks a rule I try to follow: no 'bad' data should be persisted.

For me, this includes saving data to the session object. Before we persist, we should validate, and then deal with the consequences.

To follow this rule means re-writing how `index-service` and `submit-index-service` interact with each other. Hence, most of the changes to `index-service` are to remove all the validation logic. Now, it can assume that what is in the session is valid filters.

Other changes to highlight

- we handle the page value here. The service deals with the fact the value may be undefined, a number or a string. It means downstream services like `FetchNoticesService` will get a consistent value.
- determining the page title (we mentioned we were doing this in the commit for the presenter changes)
- all mapping of notice types is dropped
- new logic to determine whether the filters section should be open of closed