Fix financial year column for billing account page (#2007)
* Fix financial year column for billing account page

https://eaflood.atlassian.net/browse/WATER-4988

While testing the billing account page, it was noticed that some rows for the financial year column were blank.
Upon further inspection, this was due to the financial year being taken from a metadata column instead of its correct column in the `bills` table.

This PR will fix the empty financial year rows for the billing account page.