Update Index validator
This update covers a suggestion made from my original review, where I said

> I think we can also avoid most of the prep work in the validator itself, and just validate the filters object directly.

What I actually meant was the payload, and with this version we are passing it pretty much straight in (the `SubmitIndexService` does clean up the notices).

We then validate the object, rather than separate validations for each element. Also, all the errors returned share the same format, so we'll see in the `SubmitIndexService` we can simplify generating valid GOV.UK UI error objects with working links.

I've added a validation for `noticeTypes` because I like to validate any 'groups' used in the UI. I know it would have to be someone maliciously creating a payload in order to send an unknown notice type. But its more about having somewhere that will blow up should we mess with the options, so we can double check the changes were intended.

Finally, I've changed the validation for `sentBy`. `FetchNoticesService` had the right idea using a `whereLike()` clause to do the searching. However, that is nullified if we then require users to enter complete email addresses.

They get more functionality if we don't require the entry be a valid email, and instead use it as provided in the `whereLike()` clause. The only constraint we put on is a max length, though that's just to prevent our QA team trying to break things!