Determine billing account backlink based on route
https://eaflood.atlassian.net/browse/WATER-4988

During testing, we noticed that the legacy billing account page could still be accessed via the internal search tool. To address this,
we added functionality to [access the new billing account page via search](https://github.com/DEFRA/water-abstraction-ui/pull/2694).

Following this change, we need to update the backlink behaviour on the billing account page. The destination of the backlink should depend
on how the user navigated to the page:

- If accessed via the internal search tool, the backlink should return the user to the search page.
- If accessed via the licence bills tab, the backlink should return the user to the bills page.

This PR updates the billing account page to handle backlink routing accordingly.