Update SubmitIndexService
This follows on the changes made to `IndexService` to move the validation into the submit service, before we save anything to the session.

The behaviour I would also expect to see when something is invalid, is that the previous results are still displayed, along with what I entered as a filter and the errors with it. I wouldn't expect the results to disappear.

Because of this, we also need to bring into this page all the logic needed to fetch results and present them ready for passing to the view.

It does mean duplication, and we should look at that once the work is complete. This page is likely to serve as a template for others so worth getting right. But for now, the user experience is paramount so the DRYness of our code will have to take a back seat.

We also see the benefit of the work to keep the references to the filters consistent (the mapping can be deleted) as well as sticking with standard Joi errors and validating the payload object (simplified validation).

The errors work as GOV.UK patterns require: a message in the summary with a working link to the control with an issue.