Start adding support for reference with Fetch Service
Every notification that gets created also gets a unique reference. It's displayed on the confirmation screen and when you click through to a notice.

In the new notice setup journey we're building it'll also be displayed on each page. We understand its a way for users to uniquely identify each notice the service sends.

However, the one place you _cannot_ view it is the Notices page! And because of that you can search for a notice using it with the filters.

This seems daft and an inconsistency, that the next few commits will resolve.

The first is to include it in FetchService. We also need to return the subtype: more on why later!