Include reference in the pageData + change type
We need to update the presenter to include the reference. But this means another column in the table, for which there isn't much room to being with.

Currently, how we display the type for each notification is one of the things taking up room. For example, 'Water abstraction alert: Warning'.

If we can shorten the notice types used in the UI, then there is plenty of space to add the reference column.

So, included in this change is (and yes I'm having to add a blasted mapping back in!) a change to how we determine the notice type to display in the UI.