Update notices batch notifications to use whole session
https://eaflood.atlassian.net/browse/WATER-5067

We previously implemented the notices journey based on the use of return periods, with the intention of adding abstraction alerts later down the line.

That time has come and we need to adjust some of the existing structures in place to allow us reuse alot of the complex logic.

To do this we will be using the session.journey to trigger any alternative logic to the returns periods.

This means passing the session object around rather than the previous attempt of restricting the session to the submit service.

When we send an abstraction alert we will need to add the threshold to the Notify payload. To simplify the setting of returns / alerts as args we will just pass the session.

This change updates the fetching the recipients and creating a notice.

This allows the flow to work up to the point of sending the emails / letters. These will fail as the Notify templates have not been added. This will be added in a later change.