Fix notice alert type validator (#2024)
https://eaflood.atlassian.net/browse/WATER-5061

When a licence monitoring station has the 'stop_or_reduce' restriction type it should be shown as an option to the user for the stop / reduce alert type. This was no happening when the only threshold was for 'stop_or_reduce'.

This change allows the restriction type 'stop_or_reduce' to be valid for all the alert types.

This can result in a user selected stop and only seeing a threshold for 'stop_or_reduce'. This is expected behaviour.