Licence tag details - user bug (#2028)
https://eaflood.atlassian.net/browse/WATER-4999

During QA a bug has been found when a user has not been assigned to the licence monitoring station record.

Assigning a user to this record is new functionality that is currently being introduced. This means the 'user' field will currently be null for most records.

This PR contains a fix to resolve the issue of the page blowing up when a user does not exist.