POC - Return lines report Discussions around the reporting contingency plans have highlighted an issue when reporting on "nil returns".
In WRLS when there is a nil return no return submission lines are stored against the return submission as the licensee has informed us that nothing has been abstracted so there is nothing to record.

However in the equivalent NALD reports, lines are recorded against a nil return with a quantity of zero if they are within the extraction period, or null if they are not.

There is apparently downstream functionality that currently relies on this report having these empty lines for nil returns. Which creates an issue for normal adhoc SQL queries since this data does not exist in the WRLS database and would not be easy or maybe even possible to try and recreate using say PL/pgSQL on our postgres DB.

As a proof of concept we are going to try and recreate this report in our service as a downloadable CSV file as this seems the method most likely to succeed. We will need to pay close attention to the performance hit running this report has on our service as it's expected to be significant.

We already have code in our system to generate CSVs. Also as part of our return logs setup process we have code that generates an array of empty return lines for a given date range. The hope is this functionality can be leveraged to assist in creating this report.