Update abstraction alert fixture
We are nearing the end of the abstraction alerts journey. We have been using the existing fixture to facilitate a simple test setup.

This has grown more complex as the journey has developed.

We previously returned an array of licence monitoring stations, we then accessed this test data using the index. This had some limitations and lead to some complex/ long code to get access to the data.

This change updates the abstraction alert fixture to export the licence monitoring stations as an object. As these have no specific name we have named them numbers, so 'one', 'two' and 'three'.

This makes it simpler to access the data.

To aid in keeping other tests unaffected we have added a optional arg to the existing function - '_licenceMonitoringStations'. This allows us to set / update the licence monitoring station data before setting getting the expected session object.

We have also introduced a simplified version of getting the relevant licence monitoring stations. These are stations that the user has selected. Once we are passed a certain point in the journey this becomes a consistent value used.

The fixture exports a new function for the relevant licence monitoring stations. This receives an array of licence refs and returns the licence monitoring stations with these licence refs assigned.

This is extremely useful to simplify test setup in the later journey and enables a better test to be performed.