Add contact to view billing account page (#2021)
* Add contact to view billing account page

https://eaflood.atlassian.net/browse/WATER-4988

While testing the change address functionality from the view billing account page, it was noticed that upon
updating the FAO(For the Attention Of), it would be displayed as part of the address on the legacy billing account page. As a result, this should
be replicated on the new billing account page.

This PR adds the FAO to the address on the view billing account page.