Refactor notices page to use payload instead of saved filters (#2039)
https://eaflood.atlassian.net/browse/WATER-4788

As part of the testing of this ticket QA noticed a couple minor issues. When entering a date, applying the filter but then trying to remove one of the date fields and apply the filter again it would correctly error but it would show the previously saved valid entry. It should show the field that was left empty.

Also when filtering using Returns invitation and any of the other values there were entries showing up that shouldn't be there. this was caused by a missing set of brackets on one of the where clauses.