Amend return version migration to exclude ended (#2047)
https://eaflood.atlassian.net/browse/WATER-4881

As part of our work in setting up quarterly returns for water companies, we need to run a one-off migration for each of them to create a new return version.

It was identified during testing that we are not handling a specific scenario.

There are a handful of cases where a licence has a return version, but the latest 'active' one ends before 1 April 2025 (when quarterly returns go live). In this case, we should not be migrating the return version record to a quarterly one.

This change excludes water company licences in this state from the migration.