Add notices abstraction alert notification personalisation (#2029)
* Add notices abstraction alert notification personalisation

https://eaflood.atlassian.net/browse/WATER-5067

We previously implemented the notices journey based on the use of return periods, with the intention of adding abstraction alerts later down the line.

That time has come and we need to adjust some of the existing structures in place to allow us reuse alot of the complex logic.

To do this we will be using the session.journey to trigger any alternative logic to the returns periods.

This means passing the session object around rather than the previous attempt of restricting the session to the submit service.

This change updates the existing presenter to create the personalisation for the notify templates.

They all expect the same data. Some are just structured slight different.

Some personalisation is missing and will be added in a later chang.

To make sure the emails are received the batch notification has been update in this change to allow the abstraction journey to use the correct presenter. This follows our existing strategy using the session.journey