Add issue email address to abstraction alerts notifications (#2053)
* Add issue email address to abstraction alerts notifications

https://eaflood.atlassian.net/browse/WATER-5067

We have previously implemented part of the notification presenter. We left out some of the personalisation to be done in later changes.

This is one of those changes.

This change adds the issuers email address to the abstraction alert notifications' personalisation. This email is used in all the templates.

This email address is set in the session from the previous page. So there is no other logic needed.